consolidating robots
